Organize internal files
