+++ client/src/index.ts
@@ -1,3 +1,7 @@
+import { buildSortFilter } from './schema/filters';
+import { Page } from './schema/pagination';
+import { Query, QueryOptions } from './schema/query';
+import { Selectable, SelectableColumn, Select } from './schema/selection';
 import { errors } from './util/errors';
 
 export interface XataRecord {
@@ -8,380 +12,6 @@ export interface XataRecord {
   read(): Promise<this>;
   update(data: Selectable<this>): Promise<this>;
   delete(): Promise<void>;
-}
-
-export type Queries<T> = {
-  [key in keyof T as T[key] extends Query<any> ? key : never]: T[key];
-};
-
-export type OmitQueries<T> = {
-  [key in keyof T as T[key] extends Query<any> ? never : key]: T[key];
-};
-
-export type OmitLinks<T> = {
-  [key in keyof T as T[key] extends XataRecord ? never : key]: T[key];
-};
-
-export type OmitMethods<T> = {
-  // eslint-disable-next-line @typescript-eslint/ban-types
-  [key in keyof T as T[key] extends Function ? never : key]: T[key];
-};
-
-export type Selectable<T> = Omit<OmitQueries<OmitMethods<T>>, 'id' | 'xata'>;
-
-type StringKeys<O> = Extract<keyof O, string>;
-type Values<O> = O[keyof O];
-
-export type SelectableColumn<O> =
-  | '*'
-  | (O extends Array<unknown>
-      ? never // TODO: Review with multiple: true
-      : O extends Record<string, any>
-      ?
-          | '*'
-          | Values<{
-              [K in StringKeys<O>]: O[K] extends Record<string, any> ? `${K}.${SelectableColumn<O[K]>}` : K;
-            }>
-      : '');
-
-type UnionToIntersection<T> = (T extends any ? (x: T) => any : never) extends (x: infer R) => any ? R : never;
-
-export type Select<T, K extends SelectableColumn<T>> = UnionToIntersection<K extends keyof T ? Pick<T, K> : T> &
-  Queries<T> &
-  XataRecord;
-
-export type Include<T> = {
-  [key in keyof T as T[key] extends XataRecord ? key : never]?: boolean | Array<keyof Selectable<T[key]>>;
-};
-
-type SortDirection = 'asc' | 'desc';
-type SortFilterExtended<T> = {
-  column: keyof T;
-  direction?: SortDirection;
-};
-
-type SortFilter<T> = SortFilterExtended<T> | keyof T;
-
-type Operator =
-  | '$gt'
-  | '$lt'
-  | '$ge'
-  | '$le'
-  | '$exists'
-  | '$notExists'
-  | '$endsWith'
-  | '$startsWith'
-  | '$pattern'
-  | '$is'
-  | '$isNot'
-  | '$contains'
-  | '$includes'
-  | '$includesSubstring'
-  | '$includesPattern'
-  | '$includesAll';
-
-// TODO: restrict constraints depending on type?
-// E.g. startsWith cannot be used with numbers
-type Constraint<T> = { [key in Operator]?: T };
-
-type DeepConstraint<T> = T extends Record<string, any>
-  ? {
-      [key in keyof T]?: T[key] | DeepConstraint<T[key]>;
-    }
-  : Constraint<T>;
-
-type ComparableType = number | Date;
-
-export const gt = <T extends ComparableType>(value: T): Constraint<T> => ({ $gt: value });
-export const ge = <T extends ComparableType>(value: T): Constraint<T> => ({ $ge: value });
-export const gte = <T extends ComparableType>(value: T): Constraint<T> => ({ $ge: value });
-export const lt = <T extends ComparableType>(value: T): Constraint<T> => ({ $lt: value });
-export const lte = <T extends ComparableType>(value: T): Constraint<T> => ({ $le: value });
-export const le = <T extends ComparableType>(value: T): Constraint<T> => ({ $le: value });
-export const exists = (column: string): Constraint<string> => ({ $exists: column });
-export const notExists = (column: string): Constraint<string> => ({ $notExists: column });
-export const startsWith = (value: string): Constraint<string> => ({ $startsWith: value });
-export const endsWith = (value: string): Constraint<string> => ({ $endsWith: value });
-export const pattern = (value: string): Constraint<string> => ({ $pattern: value });
-export const is = <T>(value: T): Constraint<T> => ({ $is: value });
-export const isNot = <T>(value: T): Constraint<T> => ({ $isNot: value });
-export const contains = <T>(value: T): Constraint<T> => ({ $contains: value });
-
-// TODO: these can only be applied to columns of type "multiple"
-export const includes = (value: string): Constraint<string> => ({ $includes: value });
-export const includesSubstring = (value: string): Constraint<string> => ({ $includesSubstring: value });
-export const includesPattern = (value: string): Constraint<string> => ({ $includesPattern: value });
-export const includesAll = (value: string): Constraint<string> => ({ $includesAll: value });
-
-type FilterConstraints<T> = {
-  [key in keyof T]?: T[key] extends Record<string, any> ? FilterConstraints<T[key]> : T[key] | DeepConstraint<T[key]>;
-};
-
-type CursorNavigationOptions = { first?: string } | { last?: string } | { after?: string; before?: string };
-type OffsetNavigationOptions = { size?: number; offset?: number };
-type PaginationOptions = CursorNavigationOptions & OffsetNavigationOptions;
-
-type BulkQueryOptions<T> = {
-  page?: PaginationOptions;
-  columns?: Array<keyof Selectable<T>>;
-  //filter?: FilterConstraints<T>;
-  sort?: SortFilter<T> | SortFilter<T>[];
-};
-
-type QueryOrConstraint<T extends XataRecord, R extends XataRecord> = Query<T, R> | Constraint<T>;
-
-type QueryMeta = { page: { cursor: string; more: boolean } };
-
-interface BasePage<T extends XataRecord, R extends XataRecord> {
-  query: Query<T, R>;
-  meta: QueryMeta;
-  records: R[];
-
-  nextPage(size?: number, offset?: number): Promise<Page<T, R>>;
-  previousPage(size?: number, offset?: number): Promise<Page<T, R>>;
-  firstPage(size?: number, offset?: number): Promise<Page<T, R>>;
-  lastPage(size?: number, offset?: number): Promise<Page<T, R>>;
-
-  hasNextPage(): boolean;
-}
-
-class Page<T extends XataRecord, R extends XataRecord> implements BasePage<T, R> {
-  readonly query: Query<T, R>;
-  readonly meta: QueryMeta;
-  readonly records: R[];
-
-  constructor(query: Query<T, R>, meta: QueryMeta, records: R[] = []) {
-    this.query = query;
-    this.meta = meta;
-    this.records = records;
-  }
-
-  async nextPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.query.getPaginated({ page: { size, offset, after: this.meta.page.cursor } });
-  }
-
-  async previousPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.query.getPaginated({ page: { size, offset, before: this.meta.page.cursor } });
-  }
-
-  async firstPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.query.getPaginated({ page: { size, offset, first: this.meta.page.cursor } });
-  }
-
-  async lastPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.query.getPaginated({ page: { size, offset, last: this.meta.page.cursor } });
-  }
-
-  // TODO: We need to add something on the backend if we want a hasPreviousPage
-  hasNextPage(): boolean {
-    return this.meta.page.more;
-  }
-}
-
-export class Query<T extends XataRecord, R extends XataRecord = T> implements BasePage<T, R> {
-  table: string;
-  repository: Repository<T>;
-
-  readonly $any?: QueryOrConstraint<T, R>[];
-  readonly $all?: QueryOrConstraint<T, R>[];
-  readonly $not?: QueryOrConstraint<T, R>[];
-  readonly $none?: QueryOrConstraint<T, R>[];
-  readonly $sort?: Record<string, SortDirection>;
-  readonly columns: SelectableColumn<T>[] = ['*'];
-
-  // Cursor pagination
-  readonly query: Query<T, R> = this;
-  readonly meta: QueryMeta = { page: { cursor: 'start', more: true } };
-  readonly records: R[] = [];
-
-  constructor(repository: Repository<T> | null, table: string, data: Partial<Query<T, R>>, parent?: Query<T, R>) {
-    if (repository) {
-      this.repository = repository;
-    } else {
-      this.repository = this as any;
-    }
-    this.table = table;
-
-    // For some reason Object.assign(this, parent) didn't work in this case
-    // so doing all this manually:
-    this.$any = parent?.$any;
-    this.$all = parent?.$all;
-    this.$not = parent?.$not;
-    this.$none = parent?.$none;
-    this.$sort = parent?.$sort;
-
-    Object.assign(this, data);
-    // These bindings are used to support deconstructing
-    // const { any, not, filter, sort } = xata.users.query()
-    this.any = this.any.bind(this);
-    this.all = this.all.bind(this);
-    this.not = this.not.bind(this);
-    this.filter = this.filter.bind(this);
-    this.sort = this.sort.bind(this);
-    this.none = this.none.bind(this);
-
-    Object.defineProperty(this, 'table', { enumerable: false });
-    Object.defineProperty(this, 'repository', { enumerable: false });
-  }
-
-  any(...queries: Query<T, R>[]): Query<T, R> {
-    return new Query<T, R>(
-      this.repository,
-      this.table,
-      {
-        $any: (this.$any || []).concat(queries)
-      },
-      this
-    );
-  }
-
-  all(...queries: Query<T, R>[]): Query<T, R> {
-    return new Query<T, R>(
-      this.repository,
-      this.table,
-      {
-        $all: (this.$all || []).concat(queries)
-      },
-      this
-    );
-  }
-
-  not(...queries: Query<T, R>[]): Query<T, R> {
-    return new Query<T, R>(
-      this.repository,
-      this.table,
-      {
-        $not: (this.$not || []).concat(queries)
-      },
-      this
-    );
-  }
-
-  none(...queries: Query<T, R>[]): Query<T, R> {
-    return new Query<T, R>(
-      this.repository,
-      this.table,
-      {
-        $none: (this.$none || []).concat(queries)
-      },
-      this
-    );
-  }
-
-  filter(constraints: FilterConstraints<T>): Query<T, R>;
-  filter<F extends keyof T>(column: F, value: FilterConstraints<T[F]> | DeepConstraint<T[F]>): Query<T, R>;
-  filter(a: any, b?: any): Query<T, R> {
-    if (arguments.length === 1) {
-      const constraints = a as FilterConstraints<T>;
-      const queries: QueryOrConstraint<T, R>[] = [];
-      for (const [column, constraint] of Object.entries(constraints)) {
-        queries.push({ [column]: constraint });
-      }
-      return new Query<T, R>(
-        this.repository,
-        this.table,
-        {
-          $all: (this.$all || []).concat(queries)
-        },
-        this
-      );
-    } else {
-      const column = a as keyof T;
-      const value = b as Partial<T[keyof T]> | Constraint<T[keyof T]>;
-      return new Query<T, R>(
-        this.repository,
-        this.table,
-        {
-          $all: (this.$all || []).concat({ [column]: value })
-        },
-        this
-      );
-    }
-  }
-
-  sort<F extends keyof T>(column: F, direction: SortDirection): Query<T, R> {
-    const sort = { ...this.$sort, [column]: direction };
-    const q = new Query<T, R>(
-      this.repository,
-      this.table,
-      {
-        $sort: sort
-      },
-      this
-    );
-
-    return q;
-  }
-
-  async getPaginated<Options extends BulkQueryOptions<T>>(
-    options: Options = {} as Options
-  ): Promise<
-    Page<T, typeof options['columns'] extends SelectableColumn<T>[] ? Select<T, typeof options['columns'][number]> : R>
-  > {
-    return this.repository._runQuery(this, options);
-  }
-
-  async *[Symbol.asyncIterator](): AsyncIterableIterator<R> {
-    for await (const [record] of this.getIterator(1)) {
-      yield record;
-    }
-  }
-
-  async *getIterator(chunk: number, options: Omit<BulkQueryOptions<T>, 'page'> = {}): AsyncGenerator<R[]> {
-    let offset = 0;
-    let end = false;
-
-    while (!end) {
-      const { records, meta } = await this.getPaginated({ ...options, page: { size: chunk, offset } });
-      yield records;
-
-      offset += chunk;
-      end = !meta.page.more;
-    }
-  }
-
-  async getMany<Options extends BulkQueryOptions<T>>(
-    options: Options = {} as Options
-  ): Promise<
-    (typeof options['columns'] extends SelectableColumn<T>[] ? Select<T, typeof options['columns'][number]> : R)[]
-  > {
-    const { records } = await this.getPaginated(options);
-    return records;
-  }
-
-  async getOne<Options extends Omit<BulkQueryOptions<T>, 'page'>>(
-    options: Options = {} as Options
-  ): Promise<
-    (typeof options['columns'] extends SelectableColumn<T>[] ? Select<T, typeof options['columns'][number]> : R) | null
-  > {
-    const records = await this.getMany({ ...options, page: { size: 1 } });
-    return records[0] || null;
-  }
-
-  async deleteAll(): Promise<number> {
-    // TODO: Return number of affected rows
-    return 0;
-  }
-
-  async nextPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.firstPage(size, offset);
-  }
-
-  async previousPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.firstPage(size, offset);
-  }
-
-  async firstPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.getPaginated({ page: { size, offset } });
-  }
-
-  async lastPage(size?: number, offset?: number): Promise<Page<T, R>> {
-    return this.getPaginated({ page: { size, offset, before: 'end' } });
-  }
-
-  hasNextPage(): boolean {
-    return this.meta.page.more;
-  }
 }
 
 export abstract class Repository<T extends XataRecord> extends Query<T> {
@@ -400,7 +30,7 @@ export abstract class Repository<T extends XataRecord> extends Query<T> {
   abstract delete(id: string): void;
 
   // Used by the Query object internally
-  abstract _runQuery<R extends XataRecord, Options extends BulkQueryOptions<T>>(
+  abstract _runQuery<R extends XataRecord, Options extends QueryOptions<T>>(
     query: Query<T, R>,
     options: Options
   ): Promise<
@@ -539,7 +169,7 @@ export class RestRepository<T extends XataRecord> extends Repository<T> {
     await this.request('DELETE', `/tables/${this.table}/data/${id}`);
   }
 
-  async _runQuery<R extends XataRecord, Options extends BulkQueryOptions<T>>(
+  async _runQuery<R extends XataRecord, Options extends QueryOptions<T>>(
     query: Query<T, R>,
     options: Options
   ): Promise<
@@ -709,23 +339,3 @@ const transformObjectLinks = (object: any) => {
     return { ...acc, [key]: value };
   }, {});
 };
-
-function buildSortFilter<T>(filter?: SortFilter<T> | SortFilter<T>[]): { [key: string]: SortDirection } | undefined {
-  if (!filter) return undefined;
-
-  const filters: SortFilter<T>[] = Array.isArray(filter) ? filter : [filter];
-
-  return filters.reduce((acc, item) => {
-    if (typeof item === 'string') {
-      return { ...acc, [item]: 'asc' };
-    } else if (isObjectSortFilter(item)) {
-      return { ...acc, [item.column]: item.direction };
-    } else {
-      return acc;
-    }
-  }, {});
-}
-
-function isObjectSortFilter<T>(filter: SortFilter<T>): filter is SortFilterExtended<T> {
-  return typeof filter === 'object' && filter.column !== undefined;
-}
